metadata: Fix so we don't require video duration to do metadata lookup

+++ src/metadata/metadata.h
@@ -441,7 +441,7 @@ void metadata_bind_albumart(struct prop *prop, rstr_t *artist, rstr_t *album);
 
 metadata_lazy_video_t *metadata_bind_video_info(rstr_t *url, rstr_t *filename,
 						rstr_t *imdb_id,
-                                                float duration,
+                                                int duration,
 						struct prop *root,
 						rstr_t *parent, int lonely,
 						int passive,
@@ -454,12 +454,12 @@ void mlv_unbind(metadata_lazy_video_t *mlv, int cleanup);
 
 void mlv_set_imdb_id(metadata_lazy_video_t *mlv, rstr_t *imdb_id);
 
-void mlv_set_duration(metadata_lazy_video_t *mlv, float duration);
+void mlv_set_duration(metadata_lazy_video_t *mlv, int duration);
 
 void mlv_set_lonely(metadata_lazy_video_t *mlv, int lonely);
 
 int mlv_direct_query(void *db, rstr_t *url, rstr_t *filename,
-                     const char *imdb_id, float duration, const char *folder,
+                     const char *imdb_id, int duration, const char *folder,
                      int lonely);
 
 /**
